+++ tests/test_core.py
@@ -413,7 +413,7 @@ def test_list_subtitles_movie(movies, mock_providers):
     assert provider_manager['opensubtitles'].plugin.list_subtitles.called
     assert provider_manager['podnapisi'].plugin.list_subtitles.called
     assert provider_manager['shooter'].plugin.list_subtitles.called
-    assert not provider_manager['subscenter'].plugin.list_subtitles.called
+    assert not provider_manager['cinemast'].plugin.list_subtitles.called
     assert provider_manager['thesubdb'].plugin.list_subtitles.called
     assert not provider_manager['tvsubtitles'].plugin.list_subtitles.called
 
@@ -435,14 +435,14 @@ def test_list_subtitles_episode(episodes, mock_providers):
     assert provider_manager['opensubtitles'].plugin.list_subtitles.called
     assert provider_manager['podnapisi'].plugin.list_subtitles.called
     assert provider_manager['shooter'].plugin.list_subtitles.called
-    assert provider_manager['subscenter'].plugin.list_subtitles.called
+    assert provider_manager['cinemast'].plugin.list_subtitles.called
     assert provider_manager['thesubdb'].plugin.list_subtitles.called
     assert provider_manager['tvsubtitles'].plugin.list_subtitles.called
 
     # test result
     assert len(subtitles) == 1
     assert sorted(subtitles[episodes['bbt_s07e05']]) == ['addic7ed', 'legendastv', 'opensubtitles', 'podnapisi',
-                                                         'shooter', 'subscenter', 'thesubdb', 'tvsubtitles']
+                                                         'shooter', 'cinemast', 'thesubdb', 'tvsubtitles']
 
 
 def test_list_subtitles_providers(episodes, mock_providers):
@@ -455,7 +455,7 @@ def test_list_subtitles_providers(episodes, mock_providers):
     assert provider_manager['addic7ed'].plugin.list_subtitles.called
     assert not provider_manager['opensubtitles'].plugin.list_subtitles.called
     assert not provider_manager['podnapisi'].plugin.list_subtitles.called
-    assert not provider_manager['subscenter'].plugin.list_subtitles.called
+    assert not provider_manager['cinemast'].plugin.list_subtitles.called
     assert not provider_manager['thesubdb'].plugin.list_subtitles.called
     assert not provider_manager['tvsubtitles'].plugin.list_subtitles.called
 
@@ -475,14 +475,14 @@ def test_list_subtitles_episode_no_hash(episodes, mock_providers):
     assert provider_manager['legendastv'].plugin.list_subtitles.called
     assert provider_manager['opensubtitles'].plugin.list_subtitles.called
     assert provider_manager['podnapisi'].plugin.list_subtitles.called
-    assert provider_manager['subscenter'].plugin.list_subtitles.called
+    assert provider_manager['cinemast'].plugin.list_subtitles.called
     assert not provider_manager['thesubdb'].plugin.list_subtitles.called
     assert provider_manager['tvsubtitles'].plugin.list_subtitles.called
 
     # test result
     assert len(subtitles) == 1
     assert sorted(subtitles[episodes['dallas_s01e03']]) == ['addic7ed', 'legendastv', 'opensubtitles', 'podnapisi',
-                                                            'shooter', 'subscenter', 'tvsubtitles']
+                                                            'shooter', 'cinemast', 'tvsubtitles']
 
 
 def test_list_subtitles_no_language(episodes, mock_providers):
@@ -496,7 +496,7 @@ def test_list_subtitles_no_language(episodes, mock_providers):
     assert not provider_manager['addic7ed'].plugin.list_subtitles.called
     assert not provider_manager['opensubtitles'].plugin.list_subtitles.called
     assert not provider_manager['podnapisi'].plugin.list_subtitles.called
-    assert not provider_manager['subscenter'].plugin.list_subtitles.called
+    assert not provider_manager['cinemast'].plugin.list_subtitles.called
     assert not provider_manager['thesubdb'].plugin.list_subtitles.called
     assert not provider_manager['tvsubtitles'].plugin.list_subtitles.called
 
@@ -521,7 +521,7 @@ def test_download_subtitles(mock_providers):
     assert not provider_manager['podnapisi'].plugin.download_subtitle.called
     assert provider_manager['thesubdb'].plugin.download_subtitle.called
     assert provider_manager['tvsubtitles'].plugin.download_subtitle.called
-    assert not provider_manager['subscenter'].plugin.download_subtitle.called
+    assert not provider_manager['cinemast'].plugin.download_subtitle.called
 
 
 @pytest.mark.integration