+++ tests/test_extensions.py
@@ -10,7 +10,7 @@ def test_registrable_extension_manager_all_extensions():
     ])
     extensions = sorted(e.name for e in manager)
     assert len(extensions) == 9
-    assert extensions == ['addic7ed', 'de7cidda', 'legendastv', 'opensubtitles', 'podnapisi', 'shooter', 'subscenter',
+    assert extensions == ['addic7ed', 'de7cidda', 'legendastv', 'opensubtitles', 'podnapisi', 'shooter', 'cinemast',
                           'thesubdb', 'tvsubtitles']
 
 
@@ -19,7 +19,7 @@ def test_registrable_extension_manager_internal_extension():
         'addic7ed = subliminal.providers.addic7ed:Addic7edProvider',
         'opensubtitles = subliminal.providers.opensubtitles:OpenSubtitlesProvider',
         'podnapisi = subliminal.providers.podnapisi:PodnapisiProvider',
-        'subscenter = subliminal.providers.subscenter:SubsCenterProvider',
+        'cinemast = subliminal.providers.cinemast:CinemastProvider',
         'thesubdb = subliminal.providers.thesubdb:TheSubDBProvider',
         'tvsubtitles = subliminal.providers.tvsubtitles:TVsubtitlesProvider'
     ])
@@ -40,7 +40,7 @@ def test_registrable_extension_manager_register():
 
 def test_registrable_extension_manager_unregister():
     manager = RegistrableExtensionManager('subliminal.test_providers', [
-        'subscenter = subliminal.providers.subscenter:SubsCenterProvider',
+        'cinemast = subliminal.providers.cinemast:CinemastProvider',
         'thesubdb = subliminal.providers.thesubdb:TheSubDBProvider',
         'tvsubtitles = subliminal.providers.tvsubtitles:TVsubtitlesProvider'
     ])
@@ -48,7 +48,7 @@ def test_registrable_extension_manager_unregister():
     manager.register('de7cidda = subliminal.providers.addic7ed:Addic7edProvider')
     manager.unregister('de7cidda = subliminal.providers.addic7ed:Addic7edProvider')
     assert len(list(manager)) == 3
-    assert set(manager.names()) == {'subscenter', 'thesubdb', 'tvsubtitles'}
+    assert set(manager.names()) == {'cinemast', 'thesubdb', 'tvsubtitles'}
 
 
 def test_provider_manager():